pyproject.toml
[project]
name = "carnot-assignment"
version = "0.1.0"
description = "Dockerized ML API with sentiment analysis and Redis caching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
]
